document public API
