+++ .gitignore
@@ -2,3 +2,4 @@
 /target
 result
 /test_data
+bacon.toml
