+++ globstar/src/context.rs
@@ -1,8 +1,9 @@
 use std::{cell::RefCell, fmt, rc::Rc};
 
-use cedar::{LocalScope, ScopeStack};
+use scope_resolution::{LocalScope, ScopeStack};
 use tree_sitter::{Node, Range, Tree};
 
+/// Pass around scope and injection data across lint rules.
 #[derive(Debug)]
 pub struct Context<'a> {
     pub(crate) root_scope: Rc<RefCell<LocalScope<'a>>>,
@@ -24,26 +25,39 @@ impl fmt::Debug for InjectedTree {
 }
 
 impl<'a> Context<'a> {
+    /// Produce the top-most scope for a given file
     pub fn root_scope(&self) -> Rc<RefCell<LocalScope<'a>>> {
         Rc::clone(&self.root_scope)
     }
+
+    /// Produce the the nearest scope that can fully contain `range`
     pub fn scope_by_range(&self, range: &Range) -> Option<Rc<RefCell<LocalScope<'a>>>> {
-        cedar::scope_by_range(self.root_scope(), range)
+        scope_resolution::scope_by_range(self.root_scope(), range)
     }
+
+    /// Produce the scope that `node` belongs to
     pub fn scope_of(&self, node: Node) -> Option<Rc<RefCell<LocalScope<'a>>>> {
         self.scope_by_range(&node.range())
     }
+
+    /// Produce a list of scopes upto the root scope, that can fully contain `range`
     pub fn scope_stack_by_range(&self, range: &Range) -> Option<ScopeStack<'a>> {
-        Some(cedar::scope_stack(self.scope_by_range(range)?))
+        Some(scope_resolution::scope_stack(self.scope_by_range(range)?))
     }
+
+    /// Produce a list of scopes upto the root scope, that `node` belongs to
     pub fn scope_stack_of(&self, node: Node) -> Option<ScopeStack<'a>> {
         self.scope_stack_by_range(&node.range())
     }
+
+    /// Produce the first injected syntax tree that is fully contained in `range`
     pub fn injected_tree_by_range(&self, range: &Range) -> Option<&InjectedTree> {
         self.injected_trees
             .iter()
             .find(|t| contains_range(&t.original_range, range))
     }
+
+    /// Produce the first injected syntax tree that is fully contained by `node`
     pub fn injected_tree_of(&self, node: &Node) -> Option<&InjectedTree> {
         self.injected_tree_by_range(&node.range())
     }
