+++ globstar/src/runner.rs
@@ -1,8 +1,8 @@
 use std::{fs, path::Path};
 
 use crate::{
-    err::{AnalysisErr, AspenErr},
-    Diagnostic, Linter, Occurrence,
+    err::{AnalysisErr, GlobstarErr},
+    Lint, Linter, Occurrence,
 };
 
 use marvin::{
@@ -15,12 +15,21 @@ use marvin::{
 use regex::RegexSet;
 
 impl Linter {
-    pub fn run_analysis(&self) -> Result<(), AspenErr> {
+    /// The whole point.
+    ///
+    /// `run_analysis` does the following:
+    /// - read environment variables required to interface with `marvin`
+    /// - load `analysis_config.json` from `$ANALYSIS_CONFIG_PATH`
+    /// - build an ignore ruleset to avoid ignored files
+    /// - build scope and injection data
+    /// - walk through the files present in `$CODE_PATH`
+    /// - store `analysis_results.json` to `$ANALYSIS_RESULT_PATH`
+    pub fn run_analysis(&self) -> Result<(), GlobstarErr> {
         let config = AnalyzerConfig::load()
             .map_err(MarvinErr::Load)
-            .map_err(AspenErr::Marvin)?;
+            .map_err(GlobstarErr::Marvin)?;
 
-        let ignore_set = RegexSet::new(&self.ignores).map_err(AspenErr::Ignore)?;
+        let ignore_set = RegexSet::new(&self.ignores).map_err(GlobstarErr::Ignore)?;
 
         let (success, _failures): (Vec<_>, Vec<_>) = config
             .files
@@ -38,7 +47,7 @@ impl Linter {
         result
             .store()
             .map_err(MarvinErr::Store)
-            .map_err(AspenErr::Marvin)
+            .map_err(GlobstarErr::Marvin)
     }
 
     fn analysis_runer_single<P: AsRef<Path>>(&self, fq_path: P) -> Result<Vec<Issue>, AnalysisErr> {
@@ -54,8 +63,9 @@ impl Linter {
             .into_iter()
             .map(
                 |Occurrence {
-                     code,
-                     diagnostic: Diagnostic { at, message },
+                     lint: Lint { code, .. },
+                     at,
+                     message,
                      ..
                  }| Issue {
                     code,
