+++ scope-resolution/Cargo.toml
@@ -1,8 +1,9 @@
 [package]
-name = "holly"
+name = "scope-resolution"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[dependencies]
+[dependencies.tree-sitter]
+git = "https://github.com/akshay-deepsource/tree-sitter"
