+++ scope-resolution/src/def.rs
