+++ scope-resolution/src/lib.rs
