+++ scope-resolution/src/scope.rs
