+++ scope-resolution/src/utils.rs
