+++ test-utils/Cargo.toml
@@ -1,9 +1,8 @@
 [package]
-name = "cedar"
+name = "test-utils"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[dependencies.tree-sitter]
-git = "https://github.com/akshay-deepsource/tree-sitter"
+[dependencies]
